+++ hyperbolix/nn_layers/hyperboloid_linear.py
@@ -1,4 +1,10 @@
-"""Hyperboloid linear layers for JAX/Flax NNX."""
+"""Hyperboloid linear layers for JAX/Flax NNX.
+
+This module contains linear transformation layers for the hyperboloid manifold,
+including the Hyperbolic Transformation Component (HTC) module from the Hypformer paper.
+
+For the core HTC/HRC functions, see hyperboloid_core module.
+"""
 
 from collections.abc import Callable
 from typing import Any
@@ -7,6 +13,8 @@ import jax
 import jax.numpy as jnp
 from flax import nnx
 from jaxtyping import Array, Float
+
+from .hyperboloid_core import htc
 
 
 class HypLinearHyperboloidFHCNN(nnx.Module):
@@ -177,3 +185,122 @@ class HypLinearHyperboloidFHCNN(nnx.Module):
             res = jnp.concatenate([res0, x_rem], axis=-1)  # (batch, out_dim)
 
         return res
+
+
+class HTCLinear(nnx.Module):
+    """Hyperbolic Transformation Component with learnable linear transformation.
+
+    This module wraps a Euclidean linear layer with the HTC operation, enabling
+    learnable transformations between hyperboloid manifolds with different curvatures.
+
+    Parameters
+    ----------
+    in_features : int
+        Input feature dimension (full hyperboloid dimension, including time component).
+    out_features : int
+        Output spatial dimension (time component is reconstructed automatically).
+    rngs : nnx.Rngs
+        Random number generators for parameter initialization.
+    use_bias : bool, optional
+        Whether to include a bias term (default: True).
+    init_bound : float, optional
+        Bound for uniform weight initialization. Weights are initialized from
+        Uniform(-init_bound, init_bound). Small values keep initial outputs
+        close to the hyperboloid origin for stable training (default: 0.02).
+    eps : float, optional
+        Small value for numerical stability (default: 1e-7).
+
+    Attributes
+    ----------
+    kernel : nnx.Param
+        Weight matrix of shape (in_features, out_features).
+    bias : nnx.Param or None
+        Bias vector of shape (out_features,) if use_bias=True, else None.
+    eps : float
+        Numerical stability parameter.
+
+    Notes
+    -----
+    Weight Initialization:
+        This layer uses small uniform initialization U(-0.02, 0.02) by default,
+        matching the initialization used by FHNN/FHCNN layers. Standard deep learning
+        initializations (Xavier, Lecun) produce weights that are too large for
+        hyperbolic operations, causing gradient explosion and training instability.
+
+    See Also
+    --------
+    hyperbolix.nn_layers.hyperboloid_core.htc : Core HTC function for functional transformations.
+    HypLinearHyperboloidFHCNN : Alternative hyperbolic linear layer with sigmoid scaling.
+
+    References
+    ----------
+    Hypformer paper (citation to be added)
+
+    Examples
+    --------
+    >>> from flax import nnx
+    >>> from hyperbolix.nn_layers import HTCLinear
+    >>> from hyperbolix.manifolds import hyperboloid
+    >>>
+    >>> # Create layer
+    >>> layer = HTCLinear(in_features=5, out_features=8, rngs=nnx.Rngs(0))
+    >>>
+    >>> # Forward pass
+    >>> x = jnp.ones((32, 5))  # batch of 32 points
+    >>> x = jax.vmap(hyperboloid.proj, in_axes=(0, None))(x, 1.0)
+    >>> y = layer(x, c_in=1.0, c_out=2.0)
+    >>> y.shape
+    (32, 9)  # 8 spatial + 1 time
+    """
+
+    def __init__(
+        self,
+        in_features: int,
+        out_features: int,
+        *,
+        rngs: nnx.Rngs,
+        use_bias: bool = True,
+        init_bound: float = 0.02,
+        eps: float = 1e-7,
+    ):
+        # Small uniform initialization for hyperbolic stability
+        # Standard initializations (Lecun, Xavier) are too large and cause gradient explosion
+        self.kernel = nnx.Param(
+            jax.random.uniform(rngs.params(), (in_features, out_features), minval=-init_bound, maxval=init_bound)
+        )
+        if use_bias:
+            self.bias = nnx.Param(jnp.zeros((out_features,)))
+        else:
+            self.bias = None
+        self.eps = eps
+
+    def __call__(
+        self,
+        x: Float[Array, "batch in_features"],
+        c_in: float = 1.0,
+        c_out: float = 1.0,
+    ) -> Float[Array, "batch out_features_plus_1"]:
+        """Apply HTC linear transformation.
+
+        Parameters
+        ----------
+        x : Array of shape (batch, in_features)
+            Input points on hyperboloid with curvature c_in.
+        c_in : float, optional
+            Input curvature (default: 1.0).
+        c_out : float, optional
+            Output curvature (default: 1.0).
+
+        Returns
+        -------
+        y : Array of shape (batch, out_features+1)
+            Output points on hyperboloid with curvature c_out.
+        """
+
+        def linear_fn(z):
+            out = z @ self.kernel[...]
+            if self.bias is not None:
+                out = out + self.bias[...]
+            return out
+
+        return htc(x, linear_fn, c_in, c_out, self.eps)
